+++ package.json
@@ -145,7 +145,7 @@
     "add": "^2.0.6",
     "aws-amplify": "4.3.44",
     "axios": "0.27.2",
-    "brokoli-ui": "1.17.2",
+    "brokoli-ui": "1.22.3",
     "chroma-js": "2.1.2",
     "copy-to-clipboard": "3.3.3",
     "ethereumjs-abi": "0.6.8",
