+++ src/routes/TraderPosition/Main/Main.styled.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const MainBox = styled('div')`
+  display: flex;
+  flex-direction: column;
+`;
