+++ src/routes/TraderPosition/TopPanel/ChainSelector/chain-options-configuration.ts
