+++ src/routes/TraderPosition/TopPanel/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { WalletConnectModal } from '../../../components/interface/WalletConnectModal/WalletConnectModal';
-import { ChainSelector } from '../ChainSelector';
-import { Nav } from '../Nav';
+import { ChainSelector } from './ChainSelector';
+import { Nav } from './Nav';
 import { PanelBox, RightContent } from './TopPanel.styled';
 
 export const TopPanel: React.FunctionComponent = React.memo(() => (
